Removes load store optimization from constant folding entirely, moves old constant folding with load store optimization to its own file as a separated optimization.
