+++ integration-tests/optimization/load-store-op.optimized.out.stdout
@@ -9,6 +9,6 @@
 1234 1234 
 1234 1234 
 1234 
-1234 1234 
+1234 1234 1234 
 1234 1234 1234 1234 1234 1234 1234 1234 1234 1234
